// src/types/models/ResponseListVOPaymentsHistoryVO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Gym Manager Swagger
 * Gym Manager REST API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { PaymentsHistoryVO } from './PaymentsHistoryVO';
import {
    PaymentsHistoryVOFromJSON,
    PaymentsHistoryVOFromJSONTyped,
    PaymentsHistoryVOToJSON,
    PaymentsHistoryVOToJSONTyped,
} from './PaymentsHistoryVO';

/**
 * 
 * @export
 * @interface ResponseListVOPaymentsHistoryVO
 */
export interface ResponseListVOPaymentsHistoryVO {
    /**
     * 현재 페이지 번호
     * @type {number}
     * @memberof ResponseListVOPaymentsHistoryVO
     */
    currentPageNo?: number;
    /**
     * 
     * @type {Array<PaymentsHistoryVO>}
     * @memberof ResponseListVOPaymentsHistoryVO
     */
    list?: Array<PaymentsHistoryVO>;
    /**
     * 페이지당 보여주는 레코드 수
     * @type {number}
     * @memberof ResponseListVOPaymentsHistoryVO
     */
    recordCountPerPage?: number;
    /**
     * 전체 개수
     * @type {number}
     * @memberof ResponseListVOPaymentsHistoryVO
     */
    totalCnt?: number;
}

/**
 * Check if a given object implements the ResponseListVOPaymentsHistoryVO interface.
 */
export function instanceOfResponseListVOPaymentsHistoryVO(value: object): value is ResponseListVOPaymentsHistoryVO {
    return true;
}

export function ResponseListVOPaymentsHistoryVOFromJSON(json: any): ResponseListVOPaymentsHistoryVO {
    return ResponseListVOPaymentsHistoryVOFromJSONTyped(json, false);
}

export function ResponseListVOPaymentsHistoryVOFromJSONTyped(json: any, ignoreDiscriminator: boolean): ResponseListVOPaymentsHistoryVO {
    if (json == null) {
        return json;
    }
    return {
        
        'currentPageNo': json['currentPageNo'] == null ? undefined : json['currentPageNo'],
        'list': json['list'] == null ? undefined : ((json['list'] as Array<any>).map(PaymentsHistoryVOFromJSON)),
        'recordCountPerPage': json['recordCountPerPage'] == null ? undefined : json['recordCountPerPage'],
        'totalCnt': json['totalCnt'] == null ? undefined : json['totalCnt'],
    };
}

export function ResponseListVOPaymentsHistoryVOToJSON(json: any): ResponseListVOPaymentsHistoryVO {
    return ResponseListVOPaymentsHistoryVOToJSONTyped(json, false);
}

export function ResponseListVOPaymentsHistoryVOToJSONTyped(value?: ResponseListVOPaymentsHistoryVO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'currentPageNo': value['currentPageNo'],
        'list': value['list'] == null ? undefined : ((value['list'] as Array<any>).map(PaymentsHistoryVOToJSON)),
        'recordCountPerPage': value['recordCountPerPage'],
        'totalCnt': value['totalCnt'],
    };
}