// src/types/models/ResponseListVOMemberVO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Gym Manager Swagger
 * Gym Manager REST API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { MemberVO } from './MemberVO';
import {
    MemberVOFromJSON,
    MemberVOFromJSONTyped,
    MemberVOToJSON,
    MemberVOToJSONTyped,
} from './MemberVO';

/**
 * 
 * @export
 * @interface ResponseListVOMemberVO
 */
export interface ResponseListVOMemberVO {
    /**
     * 현재 페이지 번호
     * @type {number}
     * @memberof ResponseListVOMemberVO
     */
    currentPageNo?: number;
    /**
     * 
     * @type {Array<MemberVO>}
     * @memberof ResponseListVOMemberVO
     */
    list?: Array<MemberVO>;
    /**
     * 페이지당 보여주는 레코드 수
     * @type {number}
     * @memberof ResponseListVOMemberVO
     */
    recordCountPerPage?: number;
    /**
     * 전체 개수
     * @type {number}
     * @memberof ResponseListVOMemberVO
     */
    totalCnt?: number;
}

/**
 * Check if a given object implements the ResponseListVOMemberVO interface.
 */
export function instanceOfResponseListVOMemberVO(value: object): value is ResponseListVOMemberVO {
    return true;
}

export function ResponseListVOMemberVOFromJSON(json: any): ResponseListVOMemberVO {
    return ResponseListVOMemberVOFromJSONTyped(json, false);
}

export function ResponseListVOMemberVOFromJSONTyped(json: any, ignoreDiscriminator: boolean): ResponseListVOMemberVO {
    if (json == null) {
        return json;
    }
    return {
        
        'currentPageNo': json['currentPageNo'] == null ? undefined : json['currentPageNo'],
        'list': json['list'] == null ? undefined : ((json['list'] as Array<any>).map(MemberVOFromJSON)),
        'recordCountPerPage': json['recordCountPerPage'] == null ? undefined : json['recordCountPerPage'],
        'totalCnt': json['totalCnt'] == null ? undefined : json['totalCnt'],
    };
}

export function ResponseListVOMemberVOToJSON(json: any): ResponseListVOMemberVO {
    return ResponseListVOMemberVOToJSONTyped(json, false);
}

export function ResponseListVOMemberVOToJSONTyped(value?: ResponseListVOMemberVO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'currentPageNo': value['currentPageNo'],
        'list': value['list'] == null ? undefined : ((value['list'] as Array<any>).map(MemberVOToJSON)),
        'recordCountPerPage': value['recordCountPerPage'],
        'totalCnt': value['totalCnt'],
    };
}