// src/types/models/ResponseListVOCalendarVO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Gym Manager Swagger
 * Gym Manager REST API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { CalendarVO } from './CalendarVO';
import {
    CalendarVOFromJSON,
    CalendarVOFromJSONTyped,
    CalendarVOToJSON,
    CalendarVOToJSONTyped,
} from './CalendarVO';

/**
 * 
 * @export
 * @interface ResponseListVOCalendarVO
 */
export interface ResponseListVOCalendarVO {
    /**
     * 현재 페이지 번호
     * @type {number}
     * @memberof ResponseListVOCalendarVO
     */
    currentPageNo?: number;
    /**
     * 
     * @type {Array<CalendarVO>}
     * @memberof ResponseListVOCalendarVO
     */
    list?: Array<CalendarVO>;
    /**
     * 페이지당 보여주는 레코드 수
     * @type {number}
     * @memberof ResponseListVOCalendarVO
     */
    recordCountPerPage?: number;
    /**
     * 전체 개수
     * @type {number}
     * @memberof ResponseListVOCalendarVO
     */
    totalCnt?: number;
}

/**
 * Check if a given object implements the ResponseListVOCalendarVO interface.
 */
export function instanceOfResponseListVOCalendarVO(value: object): value is ResponseListVOCalendarVO {
    return true;
}

export function ResponseListVOCalendarVOFromJSON(json: any): ResponseListVOCalendarVO {
    return ResponseListVOCalendarVOFromJSONTyped(json, false);
}

export function ResponseListVOCalendarVOFromJSONTyped(json: any, ignoreDiscriminator: boolean): ResponseListVOCalendarVO {
    if (json == null) {
        return json;
    }
    return {
        
        'currentPageNo': json['currentPageNo'] == null ? undefined : json['currentPageNo'],
        'list': json['list'] == null ? undefined : ((json['list'] as Array<any>).map(CalendarVOFromJSON)),
        'recordCountPerPage': json['recordCountPerPage'] == null ? undefined : json['recordCountPerPage'],
        'totalCnt': json['totalCnt'] == null ? undefined : json['totalCnt'],
    };
}

export function ResponseListVOCalendarVOToJSON(json: any): ResponseListVOCalendarVO {
    return ResponseListVOCalendarVOToJSONTyped(json, false);
}

export function ResponseListVOCalendarVOToJSONTyped(value?: ResponseListVOCalendarVO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'currentPageNo': value['currentPageNo'],
        'list': value['list'] == null ? undefined : ((value['list'] as Array<any>).map(CalendarVOToJSON)),
        'recordCountPerPage': value['recordCountPerPage'],
        'totalCnt': value['totalCnt'],
    };
}