// src/types/models/VisitChartVO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Gym Manager Swagger
 * Gym Manager REST API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface VisitChartVO
 */
export interface VisitChartVO {
    /**
     * 회원 시퀀스
     * @type {number}
     * @memberof VisitChartVO
     */
    memberSeq?: number;
    /**
     * 등록 일자
     * @type {Date}
     * @memberof VisitChartVO
     */
    regDate?: Date;
    /**
     * 방문 내역 시퀀스
     * @type {number}
     * @memberof VisitChartVO
     */
    vhSeq?: number;
    /**
     * 
     * @type {number}
     * @memberof VisitChartVO
     */
    visitCount?: number;
    /**
     * 
     * @type {string}
     * @memberof VisitChartVO
     */
    weekDay?: string;
    /**
     * 
     * @type {number}
     * @memberof VisitChartVO
     */
    weekIndex?: number;
}

/**
 * Check if a given object implements the VisitChartVO interface.
 */
export function instanceOfVisitChartVO(value: object): value is VisitChartVO {
    return true;
}

export function VisitChartVOFromJSON(json: any): VisitChartVO {
    return VisitChartVOFromJSONTyped(json, false);
}

export function VisitChartVOFromJSONTyped(json: any, ignoreDiscriminator: boolean): VisitChartVO {
    if (json == null) {
        return json;
    }
    return {
        
        'memberSeq': json['memberSeq'] == null ? undefined : json['memberSeq'],
        'regDate': json['regDate'] == null ? undefined : (new Date(json['regDate'])),
        'vhSeq': json['vhSeq'] == null ? undefined : json['vhSeq'],
        'visitCount': json['visitCount'] == null ? undefined : json['visitCount'],
        'weekDay': json['weekDay'] == null ? undefined : json['weekDay'],
        'weekIndex': json['weekIndex'] == null ? undefined : json['weekIndex'],
    };
}

export function VisitChartVOToJSON(json: any): VisitChartVO {
    return VisitChartVOToJSONTyped(json, false);
}

export function VisitChartVOToJSONTyped(value?: VisitChartVO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'memberSeq': value['memberSeq'],
        'regDate': value['regDate'] == null ? undefined : ((value['regDate']).toISOString()),
        'vhSeq': value['vhSeq'],
        'visitCount': value['visitCount'],
        'weekDay': value['weekDay'],
        'weekIndex': value['weekIndex'],
    };
}