// src/types/models/SalesChartVO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Gym Manager Swagger
 * Gym Manager REST API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface SalesChartVO
 */
export interface SalesChartVO {
    /**
     * 
     * @type {string}
     * @memberof SalesChartVO
     */
    ageGroup?: string;
    /**
     * 회원권 명
     * @type {string}
     * @memberof SalesChartVO
     */
    membershipName?: string;
    /**
     * 회원권 기간
     * @type {string}
     * @memberof SalesChartVO
     */
    membershipPeriod?: string;
    /**
     * 회원권 시퀀스
     * @type {number}
     * @memberof SalesChartVO
     */
    membershipSeq?: number;
    /**
     * 등록 일자
     * @type {Date}
     * @memberof SalesChartVO
     */
    regDate?: Date;
    /**
     * 등록자 ID
     * @type {string}
     * @memberof SalesChartVO
     */
    regId?: string;
    /**
     * 
     * @type {string}
     * @memberof SalesChartVO
     */
    regMonth?: string;
    /**
     * 
     * @type {number}
     * @memberof SalesChartVO
     */
    totalPrice?: number;
    /**
     * 사업장 시퀀스
     * @type {number}
     * @memberof SalesChartVO
     */
    wpSeq?: number;
}

/**
 * Check if a given object implements the SalesChartVO interface.
 */
export function instanceOfSalesChartVO(value: object): value is SalesChartVO {
    return true;
}

export function SalesChartVOFromJSON(json: any): SalesChartVO {
    return SalesChartVOFromJSONTyped(json, false);
}

export function SalesChartVOFromJSONTyped(json: any, ignoreDiscriminator: boolean): SalesChartVO {
    if (json == null) {
        return json;
    }
    return {
        
        'ageGroup': json['ageGroup'] == null ? undefined : json['ageGroup'],
        'membershipName': json['membershipName'] == null ? undefined : json['membershipName'],
        'membershipPeriod': json['membershipPeriod'] == null ? undefined : json['membershipPeriod'],
        'membershipSeq': json['membershipSeq'] == null ? undefined : json['membershipSeq'],
        'regDate': json['regDate'] == null ? undefined : (new Date(json['regDate'])),
        'regId': json['regId'] == null ? undefined : json['regId'],
        'regMonth': json['regMonth'] == null ? undefined : json['regMonth'],
        'totalPrice': json['totalPrice'] == null ? undefined : json['totalPrice'],
        'wpSeq': json['wpSeq'] == null ? undefined : json['wpSeq'],
    };
}

export function SalesChartVOToJSON(json: any): SalesChartVO {
    return SalesChartVOToJSONTyped(json, false);
}

export function SalesChartVOToJSONTyped(value?: SalesChartVO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'ageGroup': value['ageGroup'],
        'membershipName': value['membershipName'],
        'membershipPeriod': value['membershipPeriod'],
        'membershipSeq': value['membershipSeq'],
        'regDate': value['regDate'] == null ? undefined : ((value['regDate']).toISOString()),
        'regId': value['regId'],
        'regMonth': value['regMonth'],
        'totalPrice': value['totalPrice'],
        'wpSeq': value['wpSeq'],
    };
}