// src/types/models/PaymentsHistoryVO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Gym Manager Swagger
 * Gym Manager REST API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Membership } from './Membership';
import {
    MembershipFromJSON,
    MembershipFromJSONTyped,
    MembershipToJSON,
    MembershipToJSONTyped,
} from './Membership';

/**
 * 
 * @export
 * @interface PaymentsHistoryVO
 */
export interface PaymentsHistoryVO {
    /**
     * 회원 명
     * @type {string}
     * @memberof PaymentsHistoryVO
     */
    memberName?: string;
    /**
     * 회원 시퀀스
     * @type {number}
     * @memberof PaymentsHistoryVO
     */
    memberSeq?: number;
    /**
     * 
     * @type {Membership}
     * @memberof PaymentsHistoryVO
     */
    membership?: Membership;
    /**
     * 
     * @type {string}
     * @memberof PaymentsHistoryVO
     */
    membershipName?: string;
    /**
     * 회원권 시퀀스
     * @type {number}
     * @memberof PaymentsHistoryVO
     */
    membershipSeq?: number;
    /**
     * 결제자 명
     * @type {string}
     * @memberof PaymentsHistoryVO
     */
    phName?: string;
    /**
     * 결제 금액
     * @type {number}
     * @memberof PaymentsHistoryVO
     */
    phPrice?: number;
    /**
     * 결제내역 시퀀스
     * @type {number}
     * @memberof PaymentsHistoryVO
     */
    phSeq?: number;
    /**
     * 결제 방법
     * @type {string}
     * @memberof PaymentsHistoryVO
     */
    phType?: string;
    /**
     * 결제 일자
     * @type {Date}
     * @memberof PaymentsHistoryVO
     */
    regDate?: Date;
}

/**
 * Check if a given object implements the PaymentsHistoryVO interface.
 */
export function instanceOfPaymentsHistoryVO(value: object): value is PaymentsHistoryVO {
    return true;
}

export function PaymentsHistoryVOFromJSON(json: any): PaymentsHistoryVO {
    return PaymentsHistoryVOFromJSONTyped(json, false);
}

export function PaymentsHistoryVOFromJSONTyped(json: any, ignoreDiscriminator: boolean): PaymentsHistoryVO {
    if (json == null) {
        return json;
    }
    return {
        
        'memberName': json['memberName'] == null ? undefined : json['memberName'],
        'memberSeq': json['memberSeq'] == null ? undefined : json['memberSeq'],
        'membership': json['membership'] == null ? undefined : MembershipFromJSON(json['membership']),
        'membershipName': json['membershipName'] == null ? undefined : json['membershipName'],
        'membershipSeq': json['membershipSeq'] == null ? undefined : json['membershipSeq'],
        'phName': json['phName'] == null ? undefined : json['phName'],
        'phPrice': json['phPrice'] == null ? undefined : json['phPrice'],
        'phSeq': json['phSeq'] == null ? undefined : json['phSeq'],
        'phType': json['phType'] == null ? undefined : json['phType'],
        'regDate': json['regDate'] == null ? undefined : (new Date(json['regDate'])),
    };
}

export function PaymentsHistoryVOToJSON(json: any): PaymentsHistoryVO {
    return PaymentsHistoryVOToJSONTyped(json, false);
}

export function PaymentsHistoryVOToJSONTyped(value?: PaymentsHistoryVO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'memberName': value['memberName'],
        'memberSeq': value['memberSeq'],
        'membership': MembershipToJSON(value['membership']),
        'membershipName': value['membershipName'],
        'membershipSeq': value['membershipSeq'],
        'phName': value['phName'],
        'phPrice': value['phPrice'],
        'phSeq': value['phSeq'],
        'phType': value['phType'],
        'regDate': value['regDate'] == null ? undefined : ((value['regDate']).toISOString()),
    };
}